feat(aapd-951): use real data in web comment, move formatter to common
