+++ packages/common/src/client/appeals-api-client.js
@@ -143,7 +143,7 @@ class AppealsApiClient {
 
 	/**
 	 * @param {string} caseReference
-	 * @returns {Promise<AppealCaseWithAppellant[]>}
+	 * @returns {Promise<AppealCaseWithAppellant>}
 	 */
 	async getAppealCaseDataByCaseReference(caseReference) {
 		const endpoint = `${v2}/appeal-cases/${caseReference}`;
