+++ packages/common/src/view-model-maps/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+	...require('./appeal-headline')
+};
