+++ packages/forms-web-app/src/controllers/selected-appeal/selected-appeal.js
@@ -1,17 +1,23 @@
 const { VIEW } = require('../../lib/views');
 const { apiClient } = require('../../lib/appeals-api-client');
-const { formatAddress } = require('@pins/common/src/lib/format-address');
+const { formatHeadlineData } = require('@pins/common');
 
 /**
  * @typedef {import('appeals-service-api').Api.AppealCaseWithAppellant} AppealCaseWithAppellant
  */
 
+// Shared controller for /appeals/:caseRef and manage-appeals/:caseRef
 exports.get = async (req, res) => {
 	const appealNumber = req.params.appealNumber;
 
+	// TODO update to a new endpoint
+	// this API doesn't really satisfy the spec of AAPD-1247
+	// but I think we should move forward with this to get
+	// this code merged, then create a new endpoint and hook
+	// it up in a later iteration.
 	const caseData = await apiClient.getAppealCaseDataByCaseReference(appealNumber);
 
-	console.log(caseData);
+	console.log('👩‍🔬 ~ caseData', caseData);
 
 	const headlineData = formatHeadlineData(caseData);
 
@@ -23,56 +29,4 @@ exports.get = async (req, res) => {
 	};
 
 	res.render(VIEW.SELECTED_APPEAL.APPEAL, viewContext);
-};
-
-/**
- * @param {AppealCaseWithAppellant} caseData
- */
-const formatHeadlineData = (caseData) => {
-	const { appealTypeName, procedure, LPAApplicationReference } = caseData;
-
-	const address = formatAddress(caseData);
-
-	return [
-		{
-			key: {
-				text: 'Appeal type'
-			},
-			value: {
-				text: appealTypeName
-			}
-		},
-		{
-			key: {
-				text: 'Appeal procedure'
-			},
-			value: {
-				text: procedure
-			}
-		},
-		{
-			key: {
-				text: 'Appeal site'
-			},
-			value: {
-				text: address
-			}
-		},
-		{
-			key: {
-				text: 'Applicant'
-			},
-			value: {
-				text: 'placeholding'
-			}
-		},
-		{
-			key: {
-				text: 'Application number'
-			},
-			value: {
-				text: LPAApplicationReference
-			}
-		}
-	];
 };
