+++ packages/forms-web-app/src/routes/lpa-dashboard/your-email-address.js
@@ -18,12 +18,12 @@ const {
 
 const views = { YOUR_EMAIL_ADDRESS, ENTER_CODE };
 
-router.get('/your-email-address', getYourEmailAddress(views, true));
+router.get('/your-email-address', getYourEmailAddress(views));
 router.post(
 	'/your-email-address',
 	emailAddressValidationRules('email-address'),
 	validationErrorHandler,
-	postYourEmailAddress(views, true)
+	postYourEmailAddress(views)
 );
 
 module.exports = router;
