+++ packages/web-comment/src/routes/appeals/_appealNumber/index.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const { selectedAppeal } = require('./controller');
-const asyncHandler = require('#utils/async-handler');
+const asyncHandler = require('@pins/common/src/middleware/async-handler');
 
 const router = express.Router({ mergeParams: true });
