+++ packages/web-comment/src/routes/appeals/_appealNumber/index.njk
@@ -1,8 +1,7 @@
 {% extends "../../../index.njk" %}
 
  {% from "govuk/components/error-summary/macro.njk" import govukErrorSummary %}
- {% from "govuk/components/button/macro.njk" import govukButton %}
- {% from "govuk/components/table/macro.njk" import govukTable %}
+ {% from "govuk/components/summary-list/macro.njk" import govukSummaryList %}
 
  {% set title="
  Appeal " + appeal.status + " for comment - Comment on a planning appeal - GOV.UK" %}
@@ -23,5 +22,23 @@
        {% endif %}
 
        <h1 class="govuk-heading-l">Appeal {{appeal.status}} for comment</h1>
+
+      <style>
+				.appeal-details .govuk-summary-list__key,
+				.appeal-details .govuk-summary-list__value,
+				.appeal-details .govuk-summary-list__actions {
+					padding-top: .2rem;
+					padding-bottom: 0;
+				}
+				{# I can't tell what this is doing #}
+				.app-task-list__section {
+					padding-right: 140px
+				}
+			</style>
+
+			{{ govukSummaryList({
+				classes: "appeal-details govuk-summary-list--no-border",
+				rows: headlineData
+			}) }}
 
  {% endblock %}
