+++ packages/web-comment/src/utils/appeals-view.d.ts
@@ -4,5 +4,5 @@ export interface AppealViewModel extends Api.AppealCaseWithAppellant {
 	formattedAddress?: string;
 	formattedCaseDecisionDate?: string;
 	formattedDecisionColour?: string;
-	status?: string;
+	status?: 'open' | 'closed';
 }
